+++ packages/core/src/core/nonInteractiveToolExecutor.test.ts
@@ -5,6 +5,7 @@
  */
 
 import { describe, it, expect, vi, beforeEach } from 'vitest';
+import type { Mock } from 'vitest';
 import { executeToolCall } from './nonInteractiveToolExecutor.js';
 import type {
   ToolRegistry,
@@ -19,16 +20,18 @@ import {
   ApprovalMode,
 } from '../index.js';
 import type { Part } from '@google/genai';
-import { MockTool } from '../test-utils/tools.js';
+import { MockTool } from '../test-utils/mock-tool.js';
 
 describe('executeToolCall', () => {
   let mockToolRegistry: ToolRegistry;
   let mockTool: MockTool;
+  let executeFn: Mock;
   let abortController: AbortController;
   let mockConfig: Config;
 
   beforeEach(() => {
-    mockTool = new MockTool();
+    executeFn = vi.fn();
+    mockTool = new MockTool({ name: 'testTool', execute: executeFn });
 
     mockToolRegistry = {
       getTool: vi.fn(),
@@ -77,7 +80,7 @@ describe('executeToolCall', () => {
       returnDisplay: 'Success!',
     };
     vi.mocked(mockToolRegistry.getTool).mockReturnValue(mockTool);
-    mockTool.executeFn.mockReturnValue(toolResult);
+    executeFn.mockReturnValue(toolResult);
 
     const response = await executeToolCall(
       mockConfig,
@@ -86,7 +89,7 @@ describe('executeToolCall', () => {
     );
 
     expect(mockToolRegistry.getTool).toHaveBeenCalledWith('testTool');
-    expect(mockTool.executeFn).toHaveBeenCalledWith(request.args);
+    expect(executeFn).toHaveBeenCalledWith(request.args);
     expect(response).toStrictEqual({
       callId: 'call1',
       error: undefined,
@@ -207,7 +210,7 @@ describe('executeToolCall', () => {
       },
     };
     vi.mocked(mockToolRegistry.getTool).mockReturnValue(mockTool);
-    mockTool.executeFn.mockReturnValue(executionErrorResult);
+    executeFn.mockReturnValue(executionErrorResult);
 
     const response = await executeToolCall(
       mockConfig,
@@ -243,7 +246,7 @@ describe('executeToolCall', () => {
       prompt_id: 'prompt-id-5',
     };
     vi.mocked(mockToolRegistry.getTool).mockReturnValue(mockTool);
-    mockTool.executeFn.mockImplementation(() => {
+    executeFn.mockImplementation(() => {
       throw new Error('Something went very wrong');
     });
 
@@ -287,7 +290,7 @@ describe('executeToolCall', () => {
       returnDisplay: 'Image processed',
     };
     vi.mocked(mockToolRegistry.getTool).mockReturnValue(mockTool);
-    mockTool.executeFn.mockReturnValue(toolResult);
+    executeFn.mockReturnValue(toolResult);
 
     const response = await executeToolCall(
       mockConfig,
@@ -330,7 +333,7 @@ describe('executeToolCall', () => {
       returnDisplay: 'String returned',
     };
     vi.mocked(mockToolRegistry.getTool).mockReturnValue(mockTool);
-    mockTool.executeFn.mockReturnValue(toolResult);
+    executeFn.mockReturnValue(toolResult);
 
     const response = await executeToolCall(
       mockConfig,
@@ -358,7 +361,7 @@ describe('executeToolCall', () => {
       returnDisplay: 'Image data returned',
     };
     vi.mocked(mockToolRegistry.getTool).mockReturnValue(mockTool);
-    mockTool.executeFn.mockReturnValue(toolResult);
+    executeFn.mockReturnValue(toolResult);
 
     const response = await executeToolCall(
       mockConfig,
